feature : 일기 수정  및 삭제, 날짜기준 달력 요청 추가
DiaryCallendarDTO라는거 만들어서 ID,날짜,감정만 return시키게 해놨음.(달력용)
일기 GET할때 날짜기준, 일기 ID기준으로 둘이 분리해놨음.

일기 DiaryUpdateDTO로 사진이랑 사진외에 내용(감정, 내용) 변경 가능하게 해놓음

+++ src/main/java/Spring/MindStone/service/DiaryService/DiaryCommandServiceImpl.java
@@ -5,6 +5,7 @@ import Spring.MindStone.apiPayload.exception.handler.MemberInfoHandler;
 import Spring.MindStone.domain.diary.DailyDiary;
 import Spring.MindStone.domain.diary.DiaryImage;
 import Spring.MindStone.domain.emotion.EmotionNote;
+import Spring.MindStone.domain.enums.EmotionList;
 import Spring.MindStone.domain.member.MemberInfo;
 import Spring.MindStone.repository.DiaryRepository.DiaryImageRepository;
 import Spring.MindStone.repository.DiaryRepository.DiaryRepository;
@@ -13,6 +14,7 @@ import Spring.MindStone.service.EmotionNoteService.EmotionNoteQueryService;
 import Spring.MindStone.service.FileService;
 import Spring.MindStone.web.dto.diaryDto.DiaryResponseDTO;
 import Spring.MindStone.web.dto.diaryDto.DiarySaveDTO;
+import Spring.MindStone.web.dto.diaryDto.DiaryUpdateDTO;
 import Spring.MindStone.web.dto.diaryDto.SimpleDiaryDTO;
 import com.theokanning.openai.completion.chat.ChatCompletionRequest;
 import com.theokanning.openai.completion.chat.ChatCompletionResult;
@@ -150,24 +152,32 @@ public class DiaryCommandServiceImpl implements DiaryCommandService {
         return result.getChoices().get(0).getMessage().getContent();
     }
 
-    public SimpleDiaryDTO updateDiary(DiarySaveDTO saveDTO, Long memberId,List<MultipartFile> image){
-        MemberInfo memberInfo = memberInfoRepository.findById(memberId)
-                .orElseThrow(() -> new MemberInfoHandler(ErrorStatus.MEMBER_NOT_FOUND));
-        DailyDiary diary =diaryRepository.findDailyDiaryByDate(memberId, LocalDate.now())
+    public SimpleDiaryDTO updateDiary(DiaryUpdateDTO updateDTO, Long memberId, List<MultipartFile> image){
+        //날짜
+        DailyDiary diary =diaryRepository.findDailyDiaryByDate(memberId, updateDTO.getDate())
                 .orElseThrow(() ->new MemberInfoHandler(ErrorStatus.DIARY_NOT_FOUND));
 
-        List<DiaryImage> diaryImageList = diary.getDiaryImageList();
-
-        //1. 이미지 리스트의 aws와 레포에서 삭제
-        for(DiaryImage diaryImage : diaryImageList ){
-            fileService.deleteFile(diaryImage.getImagePath());
-            //diaryImageRepository.delete(diaryImage);
+        if(!memberId.equals(diary.getMemberInfo().getId())){
+            throw new MemberInfoHandler(ErrorStatus.DIARY_ISNT_MINE);
         }
-        //diary repo에서도 밀어버리기, orphan true라 claer하면 다 지워짐
-        diary.getDiaryImageList().clear();
 
-        //새로 바뀐 이미지 리스트 저장
-        diary.setDiaryImageList(setAwsStore(diary, image));
+        diary.update(updateDTO);//내용이나 돌 모양이 변형
+
+        //이미지의 수정 사항이 있다면 이미지 변경 진행
+        if(!image.isEmpty()){
+            List<DiaryImage> diaryImageList = diary.getDiaryImageList();
+
+            //1. 이미지 리스트의 aws와 레포에서 삭제
+            for (DiaryImage diaryImage : diaryImageList) {
+                fileService.deleteFile(diaryImage.getImagePath());
+                //diaryImageRepository.delete(diaryImage);
+            }
+            //diary repo에서도 밀어버리기, orphan true라 claer하면 다 지워짐
+            diary.getDiaryImageList().clear();
+
+            //새로 바뀐 이미지 리스트 저장
+            diary.setDiaryImageList(setAwsStore(diary, image));
+        }
 
         //바뀐 내용 반영
         diaryRepository.save(diary);
@@ -183,7 +193,7 @@ public class DiaryCommandServiceImpl implements DiaryCommandService {
         //이미지만 설정안했음!
         DailyDiary diary = DailyDiary.builder()
                 .date(saveDTO.getDate())
-                .emotion()
+                .emotion(EmotionList.fromString(saveDTO.getEmotion()))
                 .memberInfo(memberInfo)
                 .impressiveThing(saveDTO.getImpressiveThing())
                 .title(saveDTO.getTitle())
@@ -196,6 +206,31 @@ public class DiaryCommandServiceImpl implements DiaryCommandService {
 
         //일기 레포에 일기 저장
         diaryRepository.save(diary);
+
+        return new SimpleDiaryDTO(diary);
+    }
+
+    public SimpleDiaryDTO deleteDiary(Long id, Long memberId){
+        MemberInfo memberInfo = memberInfoRepository.findById(memberId)
+                .orElseThrow(() -> new MemberInfoHandler(ErrorStatus.MEMBER_NOT_FOUND));
+        DailyDiary diary =diaryRepository.findDailyDiaryByDate(memberId, LocalDate.now())
+                .orElseThrow(() ->new MemberInfoHandler(ErrorStatus.DIARY_NOT_FOUND));
+
+        if(!memberId.equals(diary.getMemberInfo().getId())){
+            throw new MemberInfoHandler(ErrorStatus.DIARY_ISNT_MINE);
+        }
+
+        List<DiaryImage> diaryImageList = diary.getDiaryImageList();
+
+        //1. 이미지 리스트의 aws와 레포에서 삭제
+        for (DiaryImage diaryImage : diaryImageList) {
+            fileService.deleteFile(diaryImage.getImagePath());
+            //diaryImageRepository.delete(diaryImage);
+        }
+        diary.getDiaryImageList().clear();
+        memberInfo.removeDiary(diary);//멤버에서도 지워줌
+
+        diaryRepository.delete(diary);
 
         return new SimpleDiaryDTO(diary);
     }
